refactor: move separate vue apps in one bigger app
