+++ index.html
@@ -36,7 +36,7 @@
   <div class="row_body">
   </div>
 </div>
-<div class="row_foot">
+<div class="row_foot" id="main_app">
   <ul id="legend_list">
     <metric-legend
       v-for="metric in metricsList"
