+++ js/main.js
@@ -5,10 +5,4 @@ import './../lib/key.js'
 import 'https://api.metricq.zih.tu-dresden.de/metricq-history-v2.js'
 
 import './uicode.js'
-import './ui/legendApp.js'
-import './ui/popupApp.js'
-import './ui/configApp.js'
-import './ui/exportApp.js'
-import './ui/presetApp.js'
-import './ui/xaxisApp.js'
-import './ui/yaxisApp.js'
+import './app.js'
